// src/shared/api/models/putApiSupplierId200.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * API
 * OpenAPI spec version: 1.0.0
 */
import type { PutApiSupplierId200Address } from './putApiSupplierId200Address';
import type { PutApiSupplierId200City } from './putApiSupplierId200City';
import type { PutApiSupplierId200ZipCode } from './putApiSupplierId200ZipCode';
import type { PutApiSupplierId200Country } from './putApiSupplierId200Country';
import type { PutApiSupplierId200Email } from './putApiSupplierId200Email';
import type { PutApiSupplierId200CountryCode } from './putApiSupplierId200CountryCode';
import type { PutApiSupplierId200Phone } from './putApiSupplierId200Phone';
import type { PutApiSupplierId200CountryISOCode } from './putApiSupplierId200CountryISOCode';
import type { PutApiSupplierId200Website } from './putApiSupplierId200Website';
import type { PutApiSupplierId200Description } from './putApiSupplierId200Description';
import type { PutApiSupplierId200CvrNumber } from './putApiSupplierId200CvrNumber';
import type { PutApiSupplierId200ContactPersonName } from './putApiSupplierId200ContactPersonName';
import type { PutApiSupplierId200ContactPersonEmail } from './putApiSupplierId200ContactPersonEmail';
import type { PutApiSupplierId200ContactPersonPhone } from './putApiSupplierId200ContactPersonPhone';
import type { PutApiSupplierId200OutletId } from './putApiSupplierId200OutletId';

export type PutApiSupplierId200 = {
    id: string;
    /**
     * @minimum -2147483648
     * @maximum 2147483647
     */
    seqId: number;
    name: string;
    address: PutApiSupplierId200Address;
    city: PutApiSupplierId200City;
    zipCode: PutApiSupplierId200ZipCode;
    country: PutApiSupplierId200Country;
    email: PutApiSupplierId200Email;
    countryCode: PutApiSupplierId200CountryCode;
    phone: PutApiSupplierId200Phone;
    countryISOCode: PutApiSupplierId200CountryISOCode;
    website: PutApiSupplierId200Website;
    description: PutApiSupplierId200Description;
    cvrNumber: PutApiSupplierId200CvrNumber;
    contactPersonName: PutApiSupplierId200ContactPersonName;
    contactPersonEmail: PutApiSupplierId200ContactPersonEmail;
    contactPersonPhone: PutApiSupplierId200ContactPersonPhone;
    outletId: PutApiSupplierId200OutletId;
    createdAt: string;
    updatedAt: string;
};