// src/shared/api/models/putApiSupplierIdBody.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * API
 * OpenAPI spec version: 1.0.0
 */

export type PutApiSupplierIdBody = {
    name?: string;
    address?: string;
    city?: string;
    zipCode?: string;
    country?: string;
    email?: string;
    countryCode?: string;
    phone?: string;
    countryISOCode?: string;
    website?: string;
    description?: string;
    cvrNumber?: string;
    contactPersonName?: string;
    contactPersonEmail?: string;
    contactPersonPhone?: string;
};